Greatly increase Text dependency upper bound
The `semigroups` package 1.12.x will not build against Text-1.1.0.0, which will be packaged with GHC 7.8 when it is released. I was able to build after increasing the upper bound with GHC 7.7.20140110 and `text-1.1.0.0`. But why increase the bound a little when you can increase it a lot? Presumably nothing `semigroups` depends on in `Text` is likely to be broken unless there's a massive API change, so I've increased the upper bound with that in mind.