Allow hashable-1.2.5 and thus GHC-7.0 (#95)
This is compomise: we still support GHC-7.0, but `Hashable Arg` instance
depends on `hashable` version.

I think this is morally ok, as *observing* difference would require
downstream to depend on `hashable`, thus they can differentiate.
But that's however not entirely true, if one e.g. uses `Arg`
as a key in `HashMap`, yet one wouldn't with old `hashable`
as then `Hashable Arg` and `Eq Arg` disagree, so it probably doesn't
work anyway.